Move utility modules to core
