+++ core/book_helpers.py
