+++ core/dispatch_fv_drop_snapshot.py
@@ -20,7 +20,7 @@ load_dotenv()
 from core.snapshot_core import format_for_display, send_bet_snapshot_to_discord
 from core.logger import get_logger
 from core.should_log_bet import MAX_POSITIVE_ODDS, MIN_NEGATIVE_ODDS
-from utils.book_helpers import parse_american_odds, filter_by_odds, ensure_side
+from core.book_helpers import parse_american_odds, filter_by_odds, ensure_side
 from core.book_whitelist import ALLOWED_BOOKS
 
 logger = get_logger(__name__)
