+++ core/dispatch_sim_only_snapshot.py
@@ -16,7 +16,7 @@ import requests
 from requests.exceptions import Timeout
 
 from core.utils import safe_load_json, post_with_retries
-from utils.book_helpers import ensure_side
+from core.book_helpers import ensure_side
 from core.logger import get_logger
 from core.market_pricer import (
     extract_best_book,
