+++ utils/snapshot_filters.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from utils.book_helpers import (
+from core.book_helpers import (
     parse_american_odds,
     filter_snapshot_rows,
 )
